Add govuk-frontend
Including govuk-frontend in application.scss is not straightforward
because url() resolution does not work for the imported SCSS.
This breaks fonts and background images. This is because when webpack
encounters the strings in those URLs, it does not see them as matching
any file it knows about and leaves them alone instead of rewriting them
to fingerprinted paths. The browser then gets 404s when it tries to
follow them.

The solution is in two parts:

1. use a SCSS function — thank you govuk-frontend developers for letting
us override the default — to make the paths inside the url() calls in
the SCSS ~-prefixed, which means they will resolve to the same paths
they would were they included from a pack file in our project.
2. tell webpacker to consider the relevant asset folders in the
govuk-frontend folder when it's trying to resolve file paths, via a
change in webpacker.yml