The first plank of an end-to-end spec for filling in form details
To run the system end to end, we will chase a single piece of data
from the view to the model and back to the view again. This commit
begins that spec with an assertion that the view is rendering properly;
future commits will add more functionality.

Make this spec a system spec, and at the same time migrate the old
feature spec to a system spec.  This is now the recommended way of
running end to end tests in Rails
(https://guides.rubyonrails.org/testing.html#system-testing). It
means we will never need the database_cleaner gem for js-enabled tests,
as Rails is now capable of running entire test inside a transaction.
This will make all our end-to-end specs faster.

We override the default driver to :rack_test (out of the box it's
selenium).