Link header error messages to relevant field
Notably, this implementation does not work for multiple validation
errors, as it only displays the first.

It also requires the mirroring setup of ids on the form groups to be
id'd manually, which seems like work.

Lastly, the gsub implementation doesn't feel like idiomatic Rails to
me - I suspect we could do this using a helper.