Support updating Personal details via the Check page
A user who clicks the "Change" link next to one of the Details on the
Check your answers page will find themselves back at
personal_details#new with a 'reviewing=true' URL parameter. This might
not be the solution in the long term but it's ok for now.

We assume we’re updating the same old PersonalDetails.last record.