Force higher lodash version
The version of lodash resolved to automatically, at the time of
writing, has a security vulnerability. We're unlikely to run afoul of
it in the near-future, but I would prefer to fix it now.

This isn't fixed upstream (in webpacker) yet, so this commit changes
the package.json config to resolve lodash to a higher version, which
does not have the vulnerability.