Add node_modules to .dockerignore
If this is not present, building the docker image on a non-linux
platform will result in a container which will not start because the Yarn
integrity check will fail.

This in turn is because the node_modules in the host are
different from the node_modules in the image.

I think this is because there are different native extensions for eg
node-sass on each platform.