Enable chromedriver specs to run in Docker
We cannot use webdrivers under Alpine Linux because the version of
Chromedriver it installs does not support that platform:

https://github.com/titusfortner/webdrivers/issues/78#issuecomment-488389351

Alpine's APK repository does provide a chromium-chromedriver package
which works. We install that using apk add in the Dockerfile.  It
depends on chromium, which we must also install using APK. I took the
list of packages from this 272-starred and updated-in-the-last-10-days
repo:

https://github.com/Zenika/alpine-chrome

I tried to get webdrivers to use the chromedriver installed via apk
but this did not work. I tried to point it to the existing binary using
the Webdrivers.install_dir option but it would always overwrite the
binary it found there with this non-Alpine-compatible version. I
experimented with the version pinning feature of webdrivers
(https://github.com/titusfortner/webdrivers/#version-pinning),
hardcoding a chromium-chromedriver version in the Dockerfile, but could
not find a way to prevent this behaviour.

So I have removed webdrivers from the Gemfile and replaced it with plain
selenium-webdriver.  This puts the onus on us to configure the
Selenium::Webdriver module with a chromedriver binary.

I did this by creating a new ENV variable called CHROMEDRIVER_PATH. This
is defined in the Dockerfile to point to the apk-installed chromedriver
binary. If no variables is present is given the spec run exits with an
exception demanding a CHROMEDRIVER_PATH. I added this error message
because otherwise Selenium hangs for many seconds waiting for
chromedriver to respond, then exits with a timeout error.