Show details of failing specs in CI checks
This commit makes it so rspec results are written to the host machine,
rather than the docker container, in our CI pipeline.

Prior to this change, this command caused the rspec-results.xml file
to be written to the docker container, which then can't be accessed
outside of that container.

A later pipeline step tries to read this file from the host machine,
but fails to parse any output because it's missing.