Use a named volume for postgres data
If we don't give postgres an explicit volume to put its data in, it
generates a volume with a long name. Getting back to a clean slate then
involves either `docker-compose down -v` or looking up the name of the
created volume using docker CLI tools.

As this has confused developers who wonder why their database is
unexpectedly persisting, make it clear that we're using a volume by
declaring one in the compose file.