Remove default entries from database config
This commit moves these 'defaults' to be set up through the docker
configuration (in environment variables) rather than as easy-to-miss
null-coalesced values in the app config.

The :test environment database name is hardcoded, because setting it via
env means that, under docker-compose, dev and test use the same
database, which is Not Good.