Set COMPOSE_FILE in azure pipeline
We used a variable called $DOCKER_OVERRIDE to invoke Docker from the
pipeline because we needed to include the docker-compose.azure.yml
configuration file.

It's possible to configure the configuration files included in the
current execution context by setting the COMPOSE_FILE ENV variable,
which removes the need for a custom override.

This is useful because it means we can turn the docker-compose calls
into make calls without having to deal with ENV inside the Makefile.