Remove Chromedriver
This was used for two integration tests where we wanted to assert on the
focus: state of a field.

- remove the Capybara config for chromedriver
- remove the section from the README
- remove the helper script to install chromedriver
- remove the section from the dockerfile

The tests which used Chromedriver to follow error summary links are cut
down to checking for the presence of the error summary. The contents of
the summary are handled by the govuk_design_system_formbuilder gem
anyway.