Mount node_modules as a volume under docker-compose
Without this change:

- running yarn install outside docker then running docker-compose up
results in a yarn error: "Integrity check: System parameters don't
match" - doing the reverse (installing in docker then running rails
server outise) results in the same error

The environment in which yarn install is run dictates the content of
node_modules, and as part of the install process, yarn creates a file
called .yarn-integrity inside node_modules.

The first entry in this .yarn-integrity file is called systemParams.

After running yarn install on Mac OS, this value is "darwin-x64-72".

After running yarn install in the Docker container, it's "linux-x64-64".

I believe this mismatch causes yarn to bail out and report the error at
the top of this commit message.

To address this, we maintain two node_modules folders. We isolate them
by mounting one as a Docker volume (named for clarity) at
app/node_modules.

This has a knock-on consequence: a brand new container will have nothing
in its node_modules folder, because the volume will have no contents.
For this reason it's necessary to run yarn install at least once for a
fresh container. I've added this to the make setup target, which is
already mandatory for fresh builds.