Do not buffer STDOUT in development
This makes logs work in Docker. Setting $stdout.sync = true means that
Ruby doesn't buffer STDOUT, which causes logs to be delivered in gobs
rather than as individual lines when running under docker-compose.

The reasons for this are rather complicated and I don't understand them,
but here is a place to begin further reading
https://stackoverflow.com/questions/39486327/stdout-being-buffered-in-docker-container