Build up initial data model
This adds 2 more models to the application, `ApplicationForm` and
`ApplicationChoice`.

We've configured those to use a cascade delete. This means that when a
Candidate is deleted from the database, we also delete their
ApplicationForms, which in turn deletes the ApplicationChoice. This is
important to keep the database clean.

It's much faster than the alternative, which is to use `dependent:
:destroy` in the models
(https://guides.rubyonrails.org/association_basics.html). This will
make Rails do the deletion itself.

In practice, we'll probably not allow deletion at all, but in test,
development and vendor sandbox this will come in handy.