Add domain model diagram to README
This uses rails-erd (https://voormedia.github.io/rails-erd) to generate
a "entity relationship diagram", also known as a domain model diagram.

Having it in the README is useful for new starters on the project. The
diagram can be manually regenerated, but by default is regenerated
whenever you run `rake db:migrate`. This hopefully will keep the
diagram and the database in sync.