Set up  CI container with vanilla make setup
Because the docker-compose build process does not respect ENV vars set
in the docker-compose.yml, there's no effect to setting
RAILS_ENV=production in this build step. We might as well set it to
test so that when we run the `rake db:setup` command in the `make
setup` step then it runs under RAILS_ENV=test, which is the behaviour
we want.