Add Rake task to generate test data
This adds a rake task to generate test data. It deletes the previously
generated data for now, to allow quick testing.

Eventually I think this class can be reused by an API endpoint that is
called by the vendors, and perhaps a vendor UI.

 Tested the result with

```
curl
"http://localhost:3000/api/v1/applications?provider=ABC&since=2019-01-01
```

And it returns the test data.