Configure erb-lint to run Rubocop on <%= %> tags
Now that a bug has been fixed in erb-lint (see fcbfae87), we can make
use of its ability to run Rubocop against embedded Ruby in views.

The configuration added to the .erb-lint file is recommended in the erb-lint
docs: https://github.com/Shopify/erb-lint#rubocop.

It includes a dependency on Rubocop, but due to compatability issues
with govuk-lint we can't yet upgrade to Rubocop 0.75.0. To fix that,
reintroduce it to the Gemfile with a < constraint.

Fix the issues erb_lint finds.