Add spec for ensuring that the correct email is sent when signing up
More specifically this is a regression test to ensure that the incorrect magic link
service is not used as it's the service that handles sending the email. E.g. if the
MagicLinkSignIn service was used after a candidate is created, tests at its current
state would have still passed but the incorrect email would have been received.