Replace email regexp with one used by Notify
We originally only checked for `/@/` because that's enough for most
cases, but this passes validations on some emails that will then fail
downstream when we send them to Notify. Notify will return a 400 which
if uncaught will lead to our app responding with a 500. This would not
be a possibility if we align our own validation with Notify's.