Add state machine for existing states
This implements a state machine for the states and transitions that we
currently have.

So, why a state machine? There are 2 main reasons:

- First, we want to make it less likely that the data goes into a weird
state. It should prevent certain impossible things from happening, like
an offer being made to a candidate who has withdrawn their application.

- Secondly, it serves as useful documentation. Having the states and
transitions in a central place makes it easy to reason about the user
journey. Most state machine libraries allow you to generate a diagrams
that give you insight in the flow.

However, there are some cases where state machines can make an
application more complicated:

- State machines often live on the ActiveRecord model with a `state`
column. Instead of that we'e choosing to have a separate
ApplicationState class. This means that we don't add many methods to
ApplicationChoice, making it less likely to become a god object. Having
a `state` attribute on ApplicationChoice makes it easier to see which
methods are added by the state machine code.

  The Workflow gem makes having state machines inside a separate class
easier than AASM (https://github.com/aasm/aasm), which is a more
frequently used state machine library. I tried to implement this using
AASM, but failed. I've used Workflow once before (about 6 years ago)
and never saw problems. I'm pretty sure we can swap it out for another
library if we so wish.

- Secondly, most state machines allow you to tack behaviour onto the
transitions. This means that the state machine eventually contains most
of the business logic. Instead of doing that, we'll keep the behaviour
in the service objects, and only use the state machine to do
transitions and to keep an eye on our states.

In terms of code, I've limited this commit to only making the current
states and transitions valid. To do this I've added the `recruited`
state, but not anything that transitions _to_ recruited. That will be
done in future PRs.