Use helpers for make API calls in request specs
Soon, all requests will have to have a API token and all POST requests
will have to send the `meta` data in the request body. To make this
easier this commit adds a helper that we can use to add defaults to the
request body and helpers. It doesn't do anything yet though.