Replace `nationalities` array with first and second
The array is a bit difficult to deal with, because it's not a `has_many`
relationship, so built in form helpers get a bit confused in what they
post to and also what values they should prepopulate individual fields
with. Moreover, I don't know if we'll save it as an array data type in the end
anyway.

Also adds some more missing accessors to `PersonalDetailsForm`.