Add the VendorApiToken model
This creates the model for the API tokens.

The tokens are saved to the database in hashed form. This means that if
we have a database leak, the tokens can't be used to use the API.

We've chosen to use the Devise token generator for the hashing of the
tokens in the database because we know it already, we're using it for
the magic link signin.