Make the offer_conditions we pass MakeAnOffer either an array or nil
We used to pass either an object { 'conditions' => ...} or nil, which was
confusing inside the object because we sometimes had to wrap it in that
structure when we passed it to the ApplicationChoice. By passing only
the value for that 'conditions' key we don't have to think about this.
It also makes validation simpler (see next commit).