Remove rejected_at db column
This is not yet implemented, though we promise it in the spec.

The date of the rejection will eventually be fetched from the audit log.
This applies to other events in the application lifecycle too. Rather
than maintain a half-implemented version here, cut it out and fake it in
the JSON response.