Align specs with the new models for choices, courses and providers
Most fixes entail creating a provider in advance of creating an
application_choice (if the provider code matters) or removing the
provider_ucas_code specified on application_choice (if it doesn't).

Extract a :for_provider_code scope on ApplicationChoice.  This is a
somewhat fiddly join and callers shouldn't know or care about it