Randomize status in the application choice factory
Having hardcoded values in factories causes tests to rely on hidden
behaviour. For example, the reject endpoint tests relied on the fact
that the default state was `application_complete` for the tests to
pass. By randomising the initial status we force the tests to be
explicit about what data they need (while still providing a non-nil
value to tests that don't care about the status).