Add the "unsubmitted" status for new application choices
This introduces the initial status for the application choices. It is
assigned in the model before creation.

To make sure the status is always filled in, we change the database
column to never allow nil-values. Note that the 4th value in the
`change_column_null` call means that existing application choices will
be `unsubmitted`.