Use autocomplete input for nationality submission
By default, the autocomplete attempts to reuse the existing `select` element for submitting values, as this plays better with existing validation systems which may not expect having to deal with arbitrary text. Because we have good validation in place, we can use the autocomplete input directly and thus avoid some bugs around users not being able to deselect a nationality once they have selected it.