In errors, namespace timestamp and attribution under meta
The fix for this is a bit of a hack to get around the way that we
construct these errors in DecisionsController: we change the
human-readable name of `timestamp` to `meta.timestamp` and do the same
for `attribution`.

I think the alternative is to create a first class Request object in the
DecisionsController which has Metadata as a nested model, and validate
that. This feels like overkill for what's effectively documentation.