Post JSON not x-www-url-form-urlencoded data in request specs
The default behaviour of a Rails request spec turns out to be that it
formencodes the incoming params, but for our API we want to accept JSON
data. Adding a Content-Type header (see below) causes the payloads not
to be correctly parsed, and blow up with schema errors like 'Cannot find
"data" key'. Explicitly transforming the Ruby hash we give params into a
string of valid JSON prevents the magical formencoding and the JSON
is delivered as-is.

The change was driven by the changed spec in this file. We wanted to
change the payload for an unconditional offer to be the same as
conditional, except the 'conditions' array would be empty. When Rails
formencoded the empty array it turned it into '[""]', which is not an
empty array, and resulted in a conditional offer being created with ""
as a condition.

Make the `get_api_request` and `post_api_request` the only places where
we build options hashes for the underlying `post` and `get` methods, and
make all specs use them (a couple were not).