Change the application_choices primary key to an integer
We used a string for this primary key so that we'd have a single
identifier for both vendor API responses (which require a 10-char
string) and internal usage. While investigating auditing approaches we
realised that having a mixture of string and integer ids inside our app
was going to be awkward.

Create a new vendor_id column which holds the old 10-char string, and
create a new id column which is an auto-incrementing integer, to use as
the internal id.

There's no need to override to_param in the presenter: ApplicationChoice
now knows how to do the right thing.