Log sent emails in development
This change makes the emails show up in the logs in development. This
is super handy because we don't have to log in to GOV.UK Notify to
inspect the contents.

Calling `deliver_now` is the standard way to send an email. We're not
entirely sure why the app previously used `deliver!`, this is possibly
because it's a method in mail-notify:

https://github.com/dxw/mail-notify/blob/5674ee1fe6b3b260592bab695df55e4b
327ed73d/lib/mail/notify/delivery_method.rb#L14-L18

Calling `deliver!` probably circumvents Rails features like logging and
instrumentation.