Do not use shoulda for a long validate_inclusion_of spec
The test description shoulda generates is an exhaustive list of
countries that fills half a screen twice during every test run: once
when the spec itself runs, and once at the bottom for the profiling
report. I think it was in the profiling report in the first place
because it checked every single entry in this long array.

This is core Rails functionality anyway. Replace the noisy shoulda spec
with a more direct handmade spec.