Avoid duplicate codes for courses and sites
By default Faker will generate random values for `Faker::Alphanumeric`
leading to a fairly high probability of duplicates when the number of
potential values is low and the number of instances generated is
reasonably high.

I noticed that `bundle exec rail generate_test_data` was failing
occasionally (maybe 50% of the time) so investigated. Faker does have a
`unique` method but this only seems to work off an in memory set of
previously generated values that gets cleared during this rake task so
I've put in a simple retry loop to make the likelihood of clashes much
lower. By resetting `Faker::Config.random` you can reproduce the issue
consistently in a test.