Test SyncProviderFromFind creates the correct resources
The e2e spec (rightly) doesn't exhaustively test all the bits and pieces
that should be created here, so spec them for the service.

Allow declaring course_code and site_code in stub_200_from_find so that
there is no magic in the spec. We might need a more flexible way of
minting fake Find API responses in future but this works for present
purposes