Remove the asset pipeline
A recent change to Sprockets caused the mere inclusion of
sprockets/rails to attempt to initialize the asset pipeline, which
depends on the presence of a manifest.js file in app/assets/config.

Remove all trace of sprockets and its brethren config.assets (which
break unless sprockets/rails is required).