Make ApplicationDate calculations respect time of day
Expiry dates such as `edit_by` and `respond_by` are generally calculated
as `(n working days after the initial_date rounded up to the end of the
say` in local time (which is always UK time for us).

I've added the `end_of_day` logic to match the requirement. Also I've
made a couple of changes to ensure things work correctly wrt timezones
(and daylight saving).

* Set the default Rails timezone to London because we want all users to see
times in UK timezone rather than the system timezone (which depends on
where the server is running).
* Use `Time.zone` rather than `Time` to create local times because the
later uses system time (which could be anywhere) rather than
'application' time (which is fixed to UK).
* Use 1st May as a `submitted_at` date in tests so that both bank
holidays and daylight savings come in to play.

There is a UX issue that I think we should address but not in the scope
of this PR. If the `edit_by` (or any time limit) is less than one day in
the future we are currently telling people time has run out. This is a
bit misleading. I would guess we would want to tell people thay have
e.g. _5 hours remaining_ or _24 minutes remaining_ which makes things a
little more complicated.