Simplify rendering of lists in summary card
Sometimes the value in a summary card is made up of multiple different
elements, like the different parts of an address. We're currently
`joining` those in the presenter with a `<br>` in between.

This means that we have to render the value as HTML, which is a
security risk. We've mitigated against that by being careful about
which things to render as HTML using the `DANGEROUS_html_value`
argument.

This commit changes the approach to avoid all sanitising and passing of
unsafe values.

I also considered the following:

```rb
<% if row[:value].is_a?(Array) %>
<%= safe_join(row[:value], '<br>'.html_safe) %>
...
```

But I don't think that is clearer.