Minor fix in ReceiveReference
Using `find_by!` causes some very odd behaviour that causes subsequent
logic to fail because it can (if the service is called more than once on
the same `application_form`) leave the `ApplicationForm#references`
collection out of step with the DB. Rather than put in an explict
`reload` I've changed the lookup by email to use `find` because we've
already validated the email address at that point and the list of
references is small.

Also renamed the initializer named arg `reference` to `feedback` because
I was confused by it.