Sync the "findable?" status from Find
We currently save all of the courses that Find returns in the database.
It turns out the Find API also returns courses that are still in draft.

As far as I can see, the `findable?` attribute on Find courses
determines whether or not it's visible to normal users. This commit
copies the `findable?` attribute to our local database, so that (later)
we can make sure we only display courses that are also in Find.

I've also considered to stop creating courses that aren't `findable?`
in the sync. This has the downside of not removing courses from the
system that flip from `findable?` to not `findable?`. We would also
have to remove the 5 draft courses that are currently in the Apply
database.