Convert summary card line breaks into HTML
This turns the line breaks a user entered into a field into HTML breaks
and paragraphs, using `simple_format`:

> Returns text transformed into HTML using simple formatting rules. Two
or more consecutive newlines(\n\n or \r\n\r\n) are considered a
paragraph and wrapped in <p> tags. One newline (\n or \r\n) is
considered a linebreak and a <br /> tag is appended. This method does
not remove the newlines from the text.

https://apidock.com/rails/ActionView/Helpers/TextHelper/simple_format