Avoid reloading references in the receive references action
We are currently reloading the application_form object in
`ReceiveReference` because of the way we set up the test data in the
specs.

By explicitly updating the `application_form.references` in the tests,
the object will already be up to date when passed into the
`ReceiveReference` and we won't have to `reload`.