Correctly reference the DFE_SIGN_IN_ISSUER ENV var
We changed this to _URI in a rebase, and didn't update Azure, which meant
ENV['DFE_SIGN_IN_URI'] was nil and caused a runtime error.

https://sentry.io/organizations/dfe-bat/issues/1336427884/?project=1765973&referrer=slack

On reflection _URI is not an improvement anyway, as this is not the only
useful URI provided by Sign-in (we'll need one for the DSI user API
too). So rather than push through the change to _URI in the Azure
pipeline, revert it to "_ISSUER" and be done.

If we find ourselves creating needing more variables to configure the
various DSI APIs then we can do some renaming later.