Make sure Sidekiq jobs report correct environment
We rely on `CUSTOM_HOSTNAME` to populate the "hosting environment" for
reporting to Sentry
(https://github.com/DFE-Digital/apply-for-postgraduate-teacher-training/
pull/478).

Because it's not set in the Sidekiq worker, they are currently
reporting development (the fallback). For an example, see
https://sentry.io/organizations/dfe-bat/issues/1341153288, which is
happening on QA.

I've also added it the the clock process, which probably needs this is
as well.