Bypass DfE Sign-in in development
Set BYPASS_DFE_SIGN_IN=true and the development environment will not send
you to DfE Sign-in for provider authentication.

The flow takes you to a form where you can fill in your desired DSI UID
and email, which will be returned to the app as though via a DfE Sign-in
authentication payload.

Note the :fake_callback method in the SessionsController which is just
an alias of :callback. This is necessary because the Omniauth
`development` strategy we're using mimics some real-world authentication
providers in that it POSTs back to our app. Because the POST has come
from "off-site", no CSRF token is expected.