Show language skills on provider application view
Create a wrapper around the SummaryListComponent

Note that the english_language_details are collected in the candidate side with a label asking for information about other languages. And the other_language_details are collected with a label asking for details about how well you speak English. ie They are the wrong way around.

That's why the test fixtures seem wrong.