Implement ProviderUser#provider
Institute a yaml file at config/provider_permissions.yml which contains
a mapping of provider codes to lists of DSI UIDs. ProviderUser#provider
gets the first (for now) matching value for its UID and returns that
provider, if it exists.

To test this end-to-end we have to be able to stub the UID that comes
back from DSI in advance, so give
DfESignInHelpers.provider_exists_in_dfe_sign_in a dfe_sign_in_uid: param
to set it.