Make the :application_choice factory friendlier
creating an application_choice without appending :single resulted in
multiple application_choices being created, because the factory depends
on application_form, and that comes with 3 application_choices out of
the box. 4 choices for 1 is a) not policy and b) confusing.

Make an explicit named trait on application_form and use it in the
application_choice factory by default