Trim down some provider_interface system specs
see_applications_spec asserted on the presence of fields in the application_form
brought along with each application_choice. Whilst it's nice to be able
to make string-against-string assertions this makes the factories more
complicated and the test harder to work with (discovered while trying to
work with it - see next commit)

provider_responds_to_application_spec copy-pasted the factories from
see_see_applications_spec, but didn't assert on the name anyway. So
those creates() can be winnowed out too.

While we're here, we prefer to have one scenario per system spec file.
In this file we had three, so improve it by combining two specs making
assertions against the same view state.