Redirect to application form when signed in
If a candidate is signed in and visit the start page, sign in page or
sign up page, they should be redirected to the application form.

This means we can simplify the ApplicationHelper, because the start
page always links to the form.

I somehow can't get this to work in the system tests - but it works
fine locally:

```
Started GET "/candidate" for ::1 at 2019-11-24 21:07:01 +0000
Processing by CandidateInterface::StartPageController#show as HTML
Redirected to http://localhost:3000/candidate/application
Filter chain halted as :redirect_to_application_if_signed_in rendered
or redirected
```