Don't crash if no site selected
This shows a validation error instead of a 500 when the user forgets to
select a site in the pick-provider screen.

Also slightly improves the validation of the validation that the user
can't choose the same course multiple times. It's now in the form
model. It's still a weird interaction though - it should probably move
to the previous page.

https://trello.com/c/Ox4W7F7G