Whitelist subject param in the URL
During a pentest we found that changing the `:subject` in a URL like:

http://localhost:3000/candidate/application/gcse/maths

into

http://localhost:3000/candidate/application/gcse/76a4sd76c

Will return a 500 and raise a Sentry error.

This restricts the values that are allowed for `:subject`, so that all
other values will return a 404 instead.

I haven't added a test because there's no good place to put it, and the
reverse is tested properly.