Only let ReceiveReference make one kind of state change
When this object can make two kinds of state transitions, the ceremony
around each transition gets lost. In this case there is a Slack
notification which happens before a choice gets `send_to_provider!`
in SendApplicationChoiceToProvider, but doesn't happen here.

The cron will pick up this change within the hour and do the transition
properly.