Add feature flag system
This feature flag system is built on top of Rollout, a gem that can do
a lot with feature flags: https://github.com/FetLife/rollout.

It's probably a bit over-engineered, but I'm hoping we can use feature
flags more in this app.

The first flag determines whether or not the pilot is open.