Put word count validation messages in locale file
Previously we weren’t passing in a custom message, and error messages read quite generically, "Reduce the word count for X".

We had some `too_long` strings in the yml but they were never being used.

Switch to a custom error message, `too_many_words`, add a default message that includes the number of words and add to becoming a teacher, subject knowledge and interview preferences.