Create GetApplicationFormsReadyToSendToProviders
This query returns all the ApplicationForms which have

- completed choices
- edit_by in the past

We used to GetApplicationChoicesReadyToSendToProvider but this operated
at the wrong level of abstraction. The choices should be sent once and
once only per form.

That class also checked whether references were present on the
application_complete choices. This is not legal and we now have an
invariant protecting us against it. The new code has confidence in
application_complete as a reliable indicator that it can be sent to the
provider.