Allow support users to impersonate candidates
This adds a button to the application form that allows a support user
to sign in as a candidate. This allows for easier manual testing by
people in the team - it avoids us from having to sign in manually.

I've enabled this only on test environments, because user impersonation
is a bit perilous when it comes to audibility and non-repudiation (we
don't want to accidentally accept an user on behalf of a real
candidate).

Part of https://trello.com/c/49yLJZEW