Generate better test applications
This adds a rake task to generate test applications.

We do this by using FactoryBot to generate a fully filled-in
application, and progressing a number of them through successive
states. This means we'll end up with 10 applications in each "stage".
There's space left at the bottom for accepting and withdrawing
applications.

It's a separate task from `GenerateTestData`, since it only generates
the applications and doesn't touch providers, courses and sites. I
suspect that we'll replace the existing test data task with this, but
I'm not sure how it's going to work out.

It doesn't have tests because they would be slow, and this is a
development-only thing for now.

In a follow up PR we'll consider how to run this on test environments.