When calculating DBD, consider withdrawal date
In the event that the last pending applicationchoice is "decided" by
being withdrawn, calculate the DBD date from the date of withdrawal.

It is not appropriate to calculate the DBD date from the date of the
last offer, as it's possible the offer was made more than 10 business
days ago. In that scenario, the offer would be declined by default on
the very next run of the sweeper.