Fix spec failure owing to inconsistent data
CourseChoicesReviewComponent's spec put it in an invalid state, where
the 'delete' link was expected on an 'application_complete' status. This
isn't possible as once the application is complete deletion is
impossible.

This bug was masked by an incomplete set of conditions in
the #withdrawable? method.

Also shrink some assertions using data- attributes.