Allow offers to be amended via the API
Once a provider makes an offer, they can amend, remove, or add
conditions as long as the candidate has not yet accepted their offer.

To allow this, we simply allow a transition between `offer` and
`offer`. This means that the API can be called any number of times, and
the offer is just updated as if it’s a new offer.

This implementation is the simplest thing that could work. However, you
could imagine that we want to treat offer amendments differently in the
future (for example, sending out a different email). I think that will
likely be discovered once we are fully live, but for now this will do.

https://trello.com/c/eMirUQxW