Include Candidate ID in API responses
This allows providers/vendors to better able to match the candidates
they received from us against existing records in their DB.

We noted an ID in our system is not guaranteed to be stable, but it's
still useful if it does happen to match data in their system. If it
doesn't then they can continue to match on other values (name, email
etc)

The ID is a string starting with “C”. This allows us to quickly see
that it’s a (C)andidate identifier, and prevents downstream services
from relying on it as an integer.

https://trello.com/c/KcV7T4zD