Make ProviderUser into an ActiveRecord model
The old behaviour remains in a LegacyProviderUser, which we will zap
when the migration is complete.

This is better because we can get by with only one "if" statement (none
required in the #provider method), and the code which is doomed is
clearly separate from the code we want to keep.