Stop rescuing mailer exceptions
The intention of this code is to raise a more specific error. Instead,
it somehow causes the exceptions to be swallowed. When a delivery error
occurs, we’re not seeing anything in the logs.

Locally the error is correctly reported to Sentry when this code is
removed.

I suspect it has something to do with the combination of backgrounded
mailers and `rescue_from`, but I’m unsure. We’ll investigate this
further.