Add UI for managing provider users
Use an array of radio buttons for want of a better UI component. While
it is possible to be associated with more than one provider, we don’t
support this via the UI, so only permit one association per
ProviderUser.

It should not be possible to create a ProviderUser without a provider,
but one day we will need to disassociate users from providers. Do not
enforce in the model that ProviderUsers have associated providers, but
do enforce it in a form object.