Start renaming references to application references
There’s a Rails method called `references` on associations:

https://apidock.com/rails/ActiveRecord/QueryMethods/references

Over the last months we’ve seen weird, unexplained, mysterious and
peculiar behaviour where associations for references aren’t reloading.
We’ve had to add a bunch of `reload` statements to the tests to make
them pass.

It could very well be unrelated to the naming collisions - another
culprit might be the default scope on references.

In any case, this renames the Reference model to ApplicationReference,
which avoids the collision and make it consistent with the other models
like ApplicationWorkExperience.

This commit does the least possible. It doesn’t rename the database
table, and doesn’t rename the factory.